Move local config into the root to make it easier to see
The local config is intended to be configured per environment. Therefore
it feels awkward to navigate deeply into the structures simply to adjust
a local configuration. Furthermore this change allows me to make an
assumption about the file right in the executed code, so if the file is
not setup properly ansible will fail even before execution!